+++ dpypelines/pipeline/shared/transforms/sdmx/compact/v20/prototype/v1.py
@@ -10,7 +10,7 @@ from dpypelines.pipeline.shared.transforms.utils import (
     pathify,
     set_key,
 )
-from dpypelines.pipeline.shared.transforms.validate_transform import (
+from dpypelines.pipeline.shared.transforms.validate_transform_v20 import (
     check_columns_of_dataframes_are_unique,
     check_header_info,
     check_header_unpacked,
